Got ncurses-intro (such as it is thus far) building.
Updated ncurses module to provide the functions ncurses-intro needs.
Broke into some ad-hoc modules.  (Next: RAII style setup/tear-down.)

Pub-ified a number of pieces of ncurses_core in order to support ncurses.

Fixed a number of constant definitions in ncurses_core (some of them
had been mistranscribed, which was harder to see before because Rust
was missing octal literals when the original port of the header file
was done).