Added screens module.  Put in the most clearly relevant methods. Would like to raii-ify the API but not sure how to handle ownership transfer in set_term (in particular, we do *not* want to run the drop method on the s argument there).
Maybe an answer could be to keep a flag in the screen, or null out its
ptr, when we are dealing with the context of set_term, which should
not deinitialize the transferred screen.  Not sure yet.

Anyway, while poking at this and trying to get the privacy right
(since I do not trust this screen code at all), I discovered what
looks like a bug in our privacy validation pass.  Will have to poke at
it more tomorrow/this week.