Added wide character string reading and writing.
Generalized the error recovery routines so that one can specify
distinct handlers for getch vs getstr.  (The other option here is to
have one handler but force the user to provide both a default raw_ch
and a str for the Return variant.  That might make more sense.)

Making a simple uniform printer required adding the HasYX trait so
that I can observe the current (y,x) coordinate of the cursor.