Quick attempt to adjust to most recent version of rustc.
rustc version: 0.9-pre (859c3ba 2013-11-28 03:51:32 -0800)

Most troubling change that I saw:

     pub enum get_err_response<'a, RET> {
         Immed(get_err_act<RET>),
-        Delay(&'static fn:Send() -> get_err_act<RET>),
+        Delay('static ||:Send -> get_err_act<RET>),
     }

A leading lifetime in front of a closure is how we are going to denote
the lifetime associated with the reference?  What?  (i.e. is this type
expression syntax really better than `&'a fn` and `&fn`?)

----

Also, I had to abandon ship on the hello{3,4,5} for now.  Probably for
the best, it seems like I was probably taking advantage of some
privacy bugs or something when I was managing to get them to compile
before.  At least it is not clear why they were able to access
e.g. `printw` (was hello3 a sibling of ncurses_core or something?)